app/requirements.txt
google-api-python-client==2.153.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
PyYAML==6.0.2
tenacity==8.5.0
python-dotenv==1.0.1
